add ServerRouter
Regarding `onRedirect`

<Redirect> in the client needs to do its work in cDM otherwise we break
React if we tried to do it in cWM.

On the server, however, we need to do the work in cWM because we don’t
ever get cDM.

I was thinking we could use the `canUseDom` to detect server v. client
to avoid creating ServerRouter, but react native can’t use dom either,
so we needed a more explicit way to know we’re on the server, so, we
have ServerRouter again.


Regarding `onMiss`

This allows folks to send 404 responses on the server.