Flatten context.router as props
We are mutating context.router to get around problems with
shouldComponentUpdate, but that also has the potential to cause subtle
bugs with user code that isn't anticipating the mutation. For example,
users should be able to use componentWillReceiveProps to detect changes
to the current route in their <Route component>s. Mutating the same
object and giving it to them makes it really difficult to detect the
change.

This commit flattens all properties of context.router in both <Route
component>s and components created using withRouter to avoid this issue.